layout/style/test/test_units_frequency.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for serialization and equivalence of frequency units</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=">Mozilla Bug </a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for serialization and equivalence of frequency units **/

/**
 * We test that for each of the following:
 *  + they reserialize to exactly what is given
 *  + if a mapping is provided, they compute to the same result as the mapping
 */
var tests = {
  "7kHz": "7000Hz",
  "300Hz": "0.3khz"
};

var p = document.getElementById("display");

for (var test in tests) {
  // We can't test this because we no longer support any properties
  // with frequency values.
  todo(false, "no tests to run, for now");
  /*
  p.setAttribute("style", "pitch: " + test);
  is(p.style.getPropertyValue("pitch"), test,
     test + " serializes to exactly itself");
  */
  // We can't test any equivalence since we don't have any properties
  // with frequency values that we compute.
/*
  var equiv = tests[test];
  if (equiv) {
    var cm1 = getComputedStyle(p, "").pitch;
    p.style.pitch = equiv;
    var cm2 = getComputedStyle(p, "").pitch;
    is(cm1, cm2, test + " should compute to the same as " + equiv);
  }
*/
}

</script>
</pre>
</body>
</html>
